rename to netutil, update go.mod
