+++ go.mod
@@ -3,8 +3,9 @@ module github.com/libp2p/go-libp2p-testing
 go 1.16
 
 require (
+	github.com/ipfs/go-log/v2 v2.4.0
 	github.com/libp2p/go-libp2p-core v0.10.0
 	github.com/multiformats/go-multiaddr v0.3.1
-	go.uber.org/atomic v1.6.0
+	go.uber.org/atomic v1.7.0
 	google.golang.org/grpc v1.31.1
 )
