+++ netutil/key_test.go
@@ -1,4 +1,4 @@
-package testutil
+package netutil
 
 import (
 	"bytes"
